Added summary view to get ratings, comments and info for recipes. Changed mapping and usage of recipeCard component. Moved create recipe to a new page.
