+++ src/routes.ts
@@ -1,6 +1,7 @@
 import type {JSX} from 'react';
 import{createElement} from 'react';
 // page components
+import CreateRecipe from './pages/CreateRecipe.tsx';
 import HomePage from './pages/HomePage.tsx';
 import NotFoundPage from './pages/NotFoundPage.tsx';
 import RecipePage from './pages/RecipePage.tsx';
@@ -15,6 +16,7 @@ interface Route {
 }
 
 export default [
+  CreateRecipe,
   HomePage,
   NotFoundPage,
   RecipePage
